Swapped alert menu for a button for permission request to start development for issue #1
Added a button on the toolbar, which defaults disabled. The button is enabled when receiving a request from the controller. When the button is clicked, permission is sent to the controller, and the button is disabled again. A text indicator in the status bar should indicate the user is requested for permission.